# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LearnerProfile {
  /** @minLength 1 */
  targetCareer: string;
  /** @items.minLength 1 */
  skills: string[];
  /** @minLength 1 */
  experience: string;
  /** @minLength 1 */
  weeklyHours: string;
  /** @minLength 1 */
  learningGoal: string;
}

export type SkillGapAnalysisInput = LearnerProfile;

export type SkillRequirementImportance = typeof SkillRequirementImportance[keyof typeof SkillRequirementImportance];


export const SkillRequirementImportance = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface SkillRequirement {
  name: string;
  importance: SkillRequirementImportance;
}

export type AcquiredSkillImportance = typeof AcquiredSkillImportance[keyof typeof AcquiredSkillImportance];


export const AcquiredSkillImportance = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface AcquiredSkill {
  name: string;
  importance: AcquiredSkillImportance;
  evidence: string;
}

export type MissingSkillImportance = typeof MissingSkillImportance[keyof typeof MissingSkillImportance];


export const MissingSkillImportance = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface MissingSkill {
  name: string;
  importance: MissingSkillImportance;
  explanation: string;
}

export interface SkillGapAnalysis {
  targetCareer: string;
  roleRequirements: SkillRequirement[];
  acquiredSkills: AcquiredSkill[];
  missingSkills: MissingSkill[];
  /**
     * @minimum 0
     * @maximum 100
     */
  matchPercentage: number;
  explanation: string;
}

export interface LearningRoadmapInput {
  profile: LearnerProfile;
  /** @minItems 1 */
  skillGaps: MissingSkill[];
}

export type RoadmapStepPriority = typeof RoadmapStepPriority[keyof typeof RoadmapStepPriority];


export const RoadmapStepPriority = {
  Critical: 'Critical',
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface RoadmapStep {
  title: string;
  focusSkill: string;
  priority: RoadmapStepPriority;
  duration: string;
  actions: string[];
  outcome: string;
}

export interface LearningRoadmap {
  targetCareer: string;
  summary: string;
  steps: RoadmapStep[];
}