# lib/api-zod/src/generated/types/acquiredSkill.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcquiredSkillImportance } from './acquiredSkillImportance';

export interface AcquiredSkill {
  name: string;
  importance: AcquiredSkillImportance;
  evidence: string;
}