# lib/api-zod/src/generated/types/learnerProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LearnerProfile {
  /** @minLength 1 */
  targetCareer: string;
  /** @items.minLength 1 */
  skills: string[];
  /** @minLength 1 */
  experience: string;
  /** @minLength 1 */
  weeklyHours: string;
  /** @minLength 1 */
  learningGoal: string;
}