# lib/api-zod/src/generated/types/learningRoadmap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapStep } from './roadmapStep';

export interface LearningRoadmap {
  targetCareer: string;
  summary: string;
  steps: RoadmapStep[];
}