# lib/api-zod/src/generated/types/learningRoadmapInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearnerProfile } from './learnerProfile';
import type { MissingSkill } from './missingSkill';

export interface LearningRoadmapInput {
  profile: LearnerProfile;
  /** @minItems 1 */
  skillGaps: MissingSkill[];
}