# lib/api-zod/src/generated/types/missingSkill.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MissingSkillImportance } from './missingSkillImportance';

export interface MissingSkill {
  name: string;
  importance: MissingSkillImportance;
  explanation: string;
}