# lib/api-zod/src/generated/types/roadmapStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapStepPriority } from './roadmapStepPriority';

export interface RoadmapStep {
  title: string;
  focusSkill: string;
  priority: RoadmapStepPriority;
  duration: string;
  actions: string[];
  outcome: string;
}