# lib/api-zod/src/generated/types/skillGapAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcquiredSkill } from './acquiredSkill';
import type { MissingSkill } from './missingSkill';
import type { SkillRequirement } from './skillRequirement';

export interface SkillGapAnalysis {
  targetCareer: string;
  roleRequirements: SkillRequirement[];
  acquiredSkills: AcquiredSkill[];
  missingSkills: MissingSkill[];
  /**
     * @minimum 0
     * @maximum 100
     */
  matchPercentage: number;
  explanation: string;
}