# lib/api-zod/src/generated/types/skillGapAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearnerProfile } from './learnerProfile';

export type SkillGapAnalysisInput = LearnerProfile;