# lib/api-zod/src/generated/types/skillRequirement.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillRequirementImportance } from './skillRequirementImportance';

export interface SkillRequirement {
  name: string;
  importance: SkillRequirementImportance;
}